refactor for further duplication and complexity reduction
